設定ダイアログの構成変更&リファクタ
